requirements.txt
alembic==0.8.10
apache-airflow==1.10.1
appnope==0.1.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.4
bleach==2.1.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.5.0
croniter==0.3.26
decorator==4.3.0
defusedxml==0.5.0
dill==0.2.8.2
docutils==0.14
Flask==0.12.4
Flask-Admin==1.4.1
Flask-AppBuilder==1.12.2
Flask-Babel==0.12.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.3.2
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gevent==1.3.7
gitdb2==2.0.5
GitPython==2.1.11
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
ipython==7.2.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.8.1
lockfile==0.12.2
lxml==4.2.5
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.1.0
monotonic==1.5
numpy==1.15.4
ordereddict==1.1
pandas==0.23.4
parso==0.3.1
password==0.2
pendulum==1.4.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psutil==4.4.2
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
python-daemon==2.1.2
python-dateutil==2.7.5
python-editor==1.0.3
python-nvd3==0.15.0
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2018.7
pytzdata==2018.7
PyYAML==3.13
requests==2.20.1
setproctitle==1.1.10
six==1.11.0
smmap2==2.0.5
SQLAlchemy==1.1.18
tabulate==0.8.2
tenacity==4.8.0
text-unidecode==1.2
thrift==0.11.0
traitlets==4.3.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
zope.deprecation==4.4.0
